Personal_Projects/PyFinanceTracker_APP/backend_v2/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
yfinance==0.2.32
pandas==2.1.3
numpy==1.26.3
pytickersymbols==1.3.2
python-dotenv==1.0.0
aiofiles==23.2.1
